MDL-58188 mod_assign: added missing require_once and updated docs
The only reason the exclusion of require_once was not exploding before
is because the call to provider_event_action() was being triggered first
which required the file. We can't rely on this happening all the time.

The doc update describes when the usage of is_event_visible() is appropriate.

Part of MDL-55611 epic.