MDL-58704 behat: Restrict search to the target tab (by css_element)
The existing search is too broad, and the text being searched exists
a number of times in the block, all them hidden, but one (the one
where we want to really check for the existence of the text).

So this patch only changes the search from block to the target
tab, doing it by css element id. Ideally we should be able to
search in tabs more naturally, but we don't have support for that right
now.