MDL-59437 behat: old "edit" cog to go to enrollments page is gone
There use to be a little cot to easily navigate from the
participations page to the enrollments page.

In boost it was already hidden, but present and the test was still
working (non-javascript).

With the conversion, the functionality is 100% gone so, right now,
the only way to navigate from participats to enrollments (that will
dissappear with with unification), is using the course admin cog.