MDL-35668 filebrowser: perf improvements for courses
When searching for non-empty children first retrieve list of all non-empty
fileareas in modules and then check access for each of them. This will be
much faster than checking access for each module and then look for files
and areas inside this module.

Do the same with course file areas