MDL-59907 behat: Fix features to be collation independent
Before the patch there were some problems with sorting
causing student1@ and student10@ to be sorted inconsistently
between databases.

Where that is not an important problem in real use, was
causing tests to fail consistently in sqlsrv.

So, adding student0@ to the tests in order to make them
stable collation-wise.