MDL-61076 core_search: Require confirmation for search areas actions
The search areas screen contains several options which are potentially
risky for a live site but can be triggered by accidentally clicking a
single button or link.

This change adds a confirmation prompt for all these options.

The enable/disable search area icon does not have a confirmation
prompt because this is nondestructive (you can just click the icon
again to change it back).