MDL-61355 SCORM: fix correct_response prefix check
The regexes to detect the optional prefixes at the start of some
interaction types are constructed to optionally match the prefix, so we
need to ensure that the matched group is really there when trying to
strip it from the given value in CRCheckValueNodes.

also remove "check for prefix" from CRcheckValueNodes

prefixes on correct responses patterns are removed in the section above
by CRremovePrefixes, so this block was redundant.
The `.pre` property for each correct_response is not used anywhere
else, so this removes their definitions too.