MDL-60915 core_dml: Change behavior to disabled by default
In stables it was decided to keep the current behavior
(not buffering/cursors) so nothing breaks/changes suddenly.

So this commit just disables the default buffer size used in
Moodle 3.5 and up (100000), defaulting to zero (feature disabled).

Still the feature can be enabled via dboptions and it's docummented
in the config-dist file.