MDL-63253 admin: Show location of the found items at search results page
The search results page showed just the titles of the found matching
pages / sections but not their location within the admin tree. In
certain cases, we even have multiple admin pages with the same title -
making it impossible to distinguish which is which.

The patch makes it so that the path of the found admin page / section is
displayed below the title.