Fixed MDL-63959 for nested dependencies.
For example the dependecy chain is the following: A->B->C
When a question (A) depends on another dependent item (B) and B hasn't
displayed (because of C's response), the $value for the B's response will be
null. In this case the can_see_item() method returned null.
Because the can_see_item() returned null (not false),
the get_pages() method displayed the question A,
because it checks for explicit false: $this->can_see_item($item) !== false